# openapi.yaml
openapi: 3.1.0
info:
  title: simpletickr API
  version: 0.1.0
  description: Simple portfolio tracker for ETFs, stocks, and crypto.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /portfolios/{portfolioId}/transactions:
    parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: [Transactions]
      operationId: recordTransaction
      summary: Record a transaction in a portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        "201":
          description: Transaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /portfolios/{portfolioId}/transactions/{id}:
    parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: [Transactions]
      operationId: amendTransaction
      summary: Amend a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        "200":
          description: Transaction amended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Transactions]
      operationId: removeTransaction
      summary: Remove a transaction
      responses:
        "204":
          description: Transaction removed
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /portfolios/{portfolioId}/transactions/trade:
    parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: [Transactions]
      operationId: recordCryptoTrade
      summary: Record a crypto-to-crypto trade (atomically records sell + buy legs)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CryptoTradeRequest"
      responses:
        "201":
          description: Trade recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoTradeResponse"
  /portfolios/{portfolioId}/transfers:
    parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Transfers]
      operationId: listTransfersForPortfolio
      summary: List transfers recorded in a portfolio
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      tags: [Transfers]
      operationId: recordTransfer
      summary: Record a transfer of an asset from one account to another, within this portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "201":
          description: Transfer recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /portfolios/{portfolioId}/transfers/{id}:
    parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: [Transfers]
      operationId: removeTransfer
      summary: Remove a transfer
      responses:
        "204":
          description: Transfer removed
        "404":
          description: Transfer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Get an account by ID
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Accounts]
      operationId: updateAccount
      summary: Update an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRequest"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Accounts]
      operationId: deleteAccount
      summary: Delete an account
      responses:
        "204":
          description: Account deleted
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Account has linked transactions and cannot be deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: List and search transactions with optional filters
      parameters:
        - name: portfolioId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TransactionType"
        - name: listingId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: assetType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AssetType"
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: accountId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Paginated list of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
...
  /transactions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Get a transaction by ID
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    AccountType:
      type: string
      enum:
        - BROKERAGE
        - CRYPTO
        - BANK
        - RETIREMENT
        - OTHER
    Account:
      type: object
      required:
        - id
        - name
        - accountType
        - transactionCount
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "My Fidelity Roth IRA"
        broker:
          type: string
          nullable: true
          example: "Fidelity"
        accountType:
          $ref: "#/components/schemas/AccountType"
        currency:
          type: string
          nullable: true
          example: "USD"
        accountNumber:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
        transactionCount:
          type: integer
          format: int64
          description: "Number of transactions linked to this account"
    AccountRequest:
      type: object
      required:
        - name
        - accountType
      properties:
        name:
          type: string
        broker:
          type: string
          nullable: true
        accountType:
          $ref: "#/components/schemas/AccountType"
        currency:
          type: string
          nullable: true
        accountNumber:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
...
    AssetType:
      type: string
      enum:
        - STOCK
        - ETF
        - CRYPTO
        - OTHER
    TransactionPage:
      type: object
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        page:
          type: integer
          description: Current page (0-indexed)
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    Transaction:
      type: object
      required:
        - id
        - portfolioId
        - listingId
        - assetId
        - type
        - quantity
        - price
        - date
        - accountId
        - account
      properties:
        id:
          type: integer
          format: int64
        portfolioId:
          type: integer
          format: int64
        listingId:
          type: integer
          format: int64
        assetId:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/TransactionType"
        quantity:
          type: number
          format: double
        price:
          type: number
          format: double
        date:
          type: string
          format: date
        fees:
          type: number
          format: double
          nullable: true
        fxRate:
          type: number
          format: double
          nullable: true
          description: "FX rate at execution time: 1 baseCurrency = fxRate listingCurrency"
        fxRateSource:
          $ref: "#/components/schemas/FxRateSource"
        accountId:
          type: integer
          format: int64
          description: "Account through which the transaction was executed"
        account:
          $ref: "#/components/schemas/Account"
        notes:
          type: string
          nullable: true
          description: "Free-form notes or context for this transaction"
        tradeId:
          type: integer
          format: int64
          nullable: true
          description: "ID of the CryptoTrade this transaction belongs to, if it is a trade leg"
    TransactionRequest:
      type: object
      required:
        - listingId
        - type
        - quantity
        - price
        - date
        - accountId
      properties:
        listingId:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/TransactionType"
        quantity:
          type: number
          format: double
        price:
          type: number
          format: double
        date:
          type: string
          format: date
        fees:
          type: number
          format: double
          nullable: true
        fxRate:
          type: number
          format: double
          nullable: true
          description: "Override the FX rate. If omitted, the server will look it up automatically."
        accountId:
          type: integer
          format: int64
          description: "Account through which the transaction was executed"
        notes:
          type: string
          nullable: true
          description: "Free-form notes or context for this transaction"
    CryptoTradeRequest:
      type: object
      required:
        - sellListingId
        - sellQuantity
        - sellPrice
        - buyListingId
        - buyQuantity
        - buyPrice
        - date
        - accountId
      properties:
        sellListingId:
          type: integer
          format: int64
        sellQuantity:
          type: number
          format: double
        sellPrice:
          type: number
          format: double
          description: "Fiat price per unit in sell listing's currency"
        buyListingId:
          type: integer
          format: int64
        buyQuantity:
          type: number
          format: double
        buyPrice:
          type: number
          format: double
          description: "Fiat price per unit in buy listing's currency"
        date:
          type: string
          format: date
        fees:
          type: number
          format: double
          nullable: true
          description: "Fees applied to the sell leg; some exchanges charge in the destination asset (future iteration)"
        accountId:
          type: integer
          format: int64
          description: "Account through which the trade was executed"
        notes:
          type: string
          nullable: true
    CryptoTradeResponse:
      type: object
      required: [id, sell, buy]
      properties:
        id:
          type: integer
          format: int64
          description: "ID of the CryptoTrade entity"
        sell:
          $ref: "#/components/schemas/Transaction"
        buy:
          $ref: "#/components/schemas/Transaction"
    Transfer:
      type: object
      description: "Moves custody of an asset between two accounts within a single portfolio. Has no price/cost-basis field — cost basis is always derived by replaying transaction/transfer history, never stored."
      required:
        - id
        - portfolioId
        - listingId
        - quantity
        - date
        - sourceAccountId
        - sourceAccount
        - destinationAccountId
        - destinationAccount
      properties:
        id:
          type: integer
          format: int64
        portfolioId:
          type: integer
          format: int64
        listingId:
          type: integer
          format: int64
        quantity:
          type: number
          format: double
        assetFeeQuantity:
          type: number
          format: double
          nullable: true
          description: "In-kind fee (e.g. crypto gas), consumed and never received anywhere. Fiat-denominated transfer fees are not modeled."
        date:
          type: string
          format: date
        sourceAccountId:
          type: integer
          format: int64
        sourceAccount:
          $ref: "#/components/schemas/Account"
        destinationAccountId:
          type: integer
          format: int64
        destinationAccount:
          $ref: "#/components/schemas/Account"
        notes:
          type: string
          nullable: true
    TransferRequest:
      type: object
      required:
        - listingId
        - quantity
        - date
        - sourceAccountId
        - destinationAccountId
      properties:
        listingId:
          type: integer
          format: int64
        quantity:
          type: number
          format: double
        assetFeeQuantity:
          type: number
          format: double
          nullable: true
          description: "In-kind fee (e.g. crypto gas), consumed and never received anywhere. Fiat-denominated transfer fees are not modeled."
        date:
          type: string
          format: date
        sourceAccountId:
          type: integer
          format: int64
        destinationAccountId:
          type: integer
          format: int64
        notes:
          type: string
          nullable: true
    TransactionType:
      type: string
      enum:
        - BUY
        - SELL
        - SPLIT
    FxRateSource:
      type: string
      nullable: true
      enum:
        - AUTO
        - USER
        - IMPORTED
...